Integrate Request evm env filling in ethapi
