+++ crates/rpc/rpc-eth-api/src/helpers/call.rs
@@ -6,7 +6,7 @@ use reth_evm::{ConfigureEvm, ConfigureEvmEnv};
 use reth_primitives::{
     revm_primitives::{
         BlockEnv, CfgEnvWithHandlerCfg, EnvWithHandlerCfg, ExecutionResult, HaltReason,
-        ResultAndState, TransactTo,
+        ResultAndState, TransactTo, TxEnv,
     },
     Bytes, TransactionSignedEcRecovered, TxKind, B256, U256,
 };
@@ -16,19 +16,22 @@ use reth_rpc_eth_types::{
     cache::db::{StateCacheDbRefMutWrapper, StateProviderTraitObjWrapper},
     error::ensure_success,
     revm_utils::{
-        apply_state_overrides, build_call_evm_env, caller_gas_allowance,
-        cap_tx_gas_limit_with_caller_allowance, get_precompiles, prepare_call_env,
+        apply_state_overrides, caller_gas_allowance, cap_tx_gas_limit_with_caller_allowance,
+        get_precompiles,
     },
     EthApiError, EthResult, RevertError, RpcInvalidTransactionError, StateCacheDb,
 };
 use reth_rpc_server_types::constants::gas_oracle::{ESTIMATE_GAS_ERROR_RATIO, MIN_TRANSACTION_GAS};
 use reth_rpc_types::{
     state::{EvmOverrides, StateOverride},
-    AccessListWithGasUsed, BlockId, Bundle, EthCallResponse, StateContext, TransactionInfo,
-    TransactionRequest,
+    AccessListWithGasUsed, BlockId, BlockOverrides, Bundle, EthCallResponse, StateContext,
+    TransactionInfo, TransactionRequest,
 };
 use revm::{Database, DatabaseCommit};
 use revm_inspectors::access_list::AccessListInspector;
+#[cfg(feature = "optimism")]
+use revm_primitives::OptimismFields;
+use std::cmp::min;
 use tracing::trace;
 
 use super::{LoadBlock, LoadPendingBlock, LoadState, LoadTransaction, SpawnBlocking, Trace};
@@ -136,7 +139,7 @@ pub trait EthCall: Call + LoadPendingBlock {
                     let state_overrides = state_override.take();
                     let overrides = EvmOverrides::new(state_overrides, block_overrides.clone());
 
-                    let env = prepare_call_env(
+                    let env = this.prepare_call_env(
                         cfg.clone(),
                         block_env.clone(),
                         tx,
@@ -206,7 +209,7 @@ pub trait EthCall: Call + LoadPendingBlock {
     {
         let state = self.state_at_block_id(at)?;
 
-        let mut env = build_call_evm_env(cfg, block, request.clone())?;
+        let mut env = Self::build_call_evm_env(cfg, block, request.clone())?;
 
         // we want to disable this in eth_createAccessList, since this is common practice used by
         // other node impls and providers <https://github.com/foundry-rs/foundry/issues/4388>
@@ -359,7 +362,7 @@ pub trait Call: LoadState + SpawnBlocking {
                 let mut db =
                     CacheDB::new(StateProviderDatabase::new(StateProviderTraitObjWrapper(&state)));
 
-                let env = prepare_call_env(
+                let env = this.prepare_call_env(
                     cfg,
                     block_env,
                     request,
@@ -530,7 +533,7 @@ pub trait Call: LoadState + SpawnBlocking {
             .unwrap_or(block_env_gas_limit);
 
         // Configure the evm env
-        let mut env = build_call_evm_env(cfg, block, request)?;
+        let mut env = Self::build_call_evm_env(cfg, block, request)?;
         let mut db = CacheDB::new(StateProviderDatabase::new(state));
 
         // Apply any state overrides if specified.
@@ -775,5 +778,429 @@ pub trait Call: LoadState + SpawnBlocking {
                 RpcInvalidTransactionError::EvmHalt(reason).into()
             }
         }
+    }
+
+    /// Creates a new [`EnvWithHandlerCfg`] to be used for executing the [`TransactionRequest`] in
+    /// `eth_call`.
+    ///
+    /// Note: this does _not_ access the Database to check the sender.
+    fn build_call_evm_env(
+        cfg: CfgEnvWithHandlerCfg,
+        block: BlockEnv,
+        request: TransactionRequest,
+    ) -> EthResult<EnvWithHandlerCfg> {
+        /// Configures a new [`TxEnv`]  for the [`TransactionRequest`]
+        ///
+        /// All [`TxEnv`] fields are derived from the given [`TransactionRequest`], if fields are
+        /// `None`, they fall back to the [`BlockEnv`]'s settings.
+        fn create_txn_env(block_env: &BlockEnv, request: TransactionRequest) -> EthResult<TxEnv> {
+            // Ensure that if versioned hashes are set, they're not empty
+            if request.blob_versioned_hashes.as_ref().map_or(false, |hashes| hashes.is_empty()) {
+                return Err(RpcInvalidTransactionError::BlobTransactionMissingBlobHashes.into())
+            }
+
+            let TransactionRequest {
+                from,
+                to,
+                gas_price,
+                max_fee_per_gas,
+                max_priority_fee_per_gas,
+                gas,
+                value,
+                input,
+                nonce,
+                access_list,
+                chain_id,
+                blob_versioned_hashes,
+                max_fee_per_blob_gas,
+                ..
+            } = request;
+
+            let CallFees { max_priority_fee_per_gas, gas_price, max_fee_per_blob_gas } =
+                CallFees::ensure_fees(
+                    gas_price.map(U256::from),
+                    max_fee_per_gas.map(U256::from),
+                    max_priority_fee_per_gas.map(U256::from),
+                    block_env.basefee,
+                    blob_versioned_hashes.as_deref(),
+                    max_fee_per_blob_gas.map(U256::from),
+                    block_env.get_blob_gasprice().map(U256::from),
+                )?;
+
+            let gas_limit =
+                gas.unwrap_or_else(|| block_env.gas_limit.min(U256::from(u64::MAX)).to());
+            let env = TxEnv {
+                gas_limit: gas_limit
+                    .try_into()
+                    .map_err(|_| RpcInvalidTransactionError::GasUintOverflow)?,
+                nonce,
+                caller: from.unwrap_or_default(),
+                gas_price,
+                gas_priority_fee: max_priority_fee_per_gas,
+                transact_to: to.unwrap_or(TxKind::Create),
+                value: value.unwrap_or_default(),
+                data: input.try_into_unique_input()?.unwrap_or_default(),
+                chain_id,
+                access_list: access_list
+                    .map(reth_rpc_types::AccessList::into_flattened)
+                    .unwrap_or_default(),
+                // EIP-4844 fields
+                blob_hashes: blob_versioned_hashes.unwrap_or_default(),
+                max_fee_per_blob_gas,
+                #[cfg(feature = "optimism")]
+                optimism: OptimismFields { enveloped_tx: Some(Bytes::new()), ..Default::default() },
+            };
+
+            Ok(env)
+        }
+        let tx = create_txn_env(&block, request)?;
+        Ok(EnvWithHandlerCfg::new_with_cfg_env(cfg, block, tx))
+    }
+
+    /// Prepares the [`EnvWithHandlerCfg`] for execution.
+    ///
+    /// Does not commit any changes to the underlying database.
+    ///
+    /// EVM settings:
+    ///  - `disable_block_gas_limit` is set to `true`
+    ///  - `disable_eip3607` is set to `true`
+    ///  - `disable_base_fee` is set to `true`
+    ///  - `nonce` is set to `None`
+    fn prepare_call_env<DB>(
+        &self,
+        mut cfg: CfgEnvWithHandlerCfg,
+        mut block: BlockEnv,
+        request: TransactionRequest,
+        gas_limit: u64,
+        db: &mut CacheDB<DB>,
+        overrides: EvmOverrides,
+    ) -> EthResult<EnvWithHandlerCfg>
+    where
+        DB: DatabaseRef,
+        EthApiError: From<<DB as DatabaseRef>::Error>,
+    {
+        /// Applies the given block overrides to the env
+        fn apply_block_overrides(overrides: BlockOverrides, env: &mut BlockEnv) {
+            let BlockOverrides {
+                number,
+                difficulty,
+                time,
+                gas_limit,
+                coinbase,
+                random,
+                base_fee,
+                block_hash: _,
+            } = overrides;
+
+            if let Some(number) = number {
+                env.number = number;
+            }
+            if let Some(difficulty) = difficulty {
+                env.difficulty = difficulty;
+            }
+            if let Some(time) = time {
+                env.timestamp = U256::from(time);
+            }
+            if let Some(gas_limit) = gas_limit {
+                env.gas_limit = U256::from(gas_limit);
+            }
+            if let Some(coinbase) = coinbase {
+                env.coinbase = coinbase;
+            }
+            if let Some(random) = random {
+                env.prevrandao = Some(random);
+            }
+            if let Some(base_fee) = base_fee {
+                env.basefee = base_fee;
+            }
+        }
+        // we want to disable this in eth_call, since this is common practice used by other node
+        // impls and providers <https://github.com/foundry-rs/foundry/issues/4388>
+        cfg.disable_block_gas_limit = true;
+
+        // Disabled because eth_call is sometimes used with eoa senders
+        // See <https://github.com/paradigmxyz/reth/issues/1959>
+        cfg.disable_eip3607 = true;
+
+        // The basefee should be ignored for eth_call
+        // See:
+        // <https://github.com/ethereum/go-ethereum/blob/ee8e83fa5f6cb261dad2ed0a7bbcde4930c41e6c/internal/ethapi/api.go#L985>
+        cfg.disable_base_fee = true;
+
+        // apply block overrides, we need to apply them first so that they take effect when we we
+        // create the evm env via `build_call_evm_env`, e.g. basefee
+        if let Some(mut block_overrides) = overrides.block {
+            if let Some(block_hashes) = block_overrides.block_hash.take() {
+                // override block hashes
+                db.block_hashes
+                    .extend(block_hashes.into_iter().map(|(num, hash)| (U256::from(num), hash)))
+            }
+            apply_block_overrides(*block_overrides, &mut block);
+        }
+
+        let request_gas = request.gas;
+        let mut env = Self::build_call_evm_env(cfg, block, request)?;
+        // set nonce to None so that the next nonce is used when transacting the call
+        env.tx.nonce = None;
+
+        // apply state overrides
+        if let Some(state_overrides) = overrides.state {
+            apply_state_overrides(state_overrides, db)?;
+        }
+
+        if request_gas.is_none() {
+            // No gas limit was provided in the request, so we need to cap the transaction gas limit
+            if env.tx.gas_price > U256::ZERO {
+                // If gas price is specified, cap transaction gas limit with caller allowance
+                trace!(target: "rpc::eth::call", ?env, "Applying gas limit cap with caller allowance");
+                cap_tx_gas_limit_with_caller_allowance(db, &mut env.tx)?;
+            } else {
+                // If no gas price is specified, use maximum allowed gas limit. The reason for this
+                // is that both Erigon and Geth use pre-configured gas cap even if
+                // it's possible to derive the gas limit from the block:
+                // <https://github.com/ledgerwatch/erigon/blob/eae2d9a79cb70dbe30b3a6b79c436872e4605458/cmd/rpcdaemon/commands/trace_adhoc.go#L956
+                // https://github.com/ledgerwatch/erigon/blob/eae2d9a79cb70dbe30b3a6b79c436872e4605458/eth/ethconfig/config.go#L94>
+                trace!(target: "rpc::eth::call", ?env, "Applying gas limit cap as the maximum gas limit");
+                env.tx.gas_limit = gas_limit;
+            }
+        }
+
+        Ok(env)
+    }
+}
+
+/// Helper type for representing the fees of a [`TransactionRequest`]
+#[derive(Debug)]
+pub struct CallFees {
+    /// EIP-1559 priority fee
+    max_priority_fee_per_gas: Option<U256>,
+    /// Unified gas price setting
+    ///
+    /// Will be the configured `basefee` if unset in the request
+    ///
+    /// `gasPrice` for legacy,
+    /// `maxFeePerGas` for EIP-1559
+    gas_price: U256,
+    /// Max Fee per Blob gas for EIP-4844 transactions
+    max_fee_per_blob_gas: Option<U256>,
+}
+
+// === impl CallFees ===
+
+impl CallFees {
+    /// Ensures the fields of a [`TransactionRequest`] are not conflicting.
+    ///
+    /// # EIP-4844 transactions
+    ///
+    /// Blob transactions have an additional fee parameter `maxFeePerBlobGas`.
+    /// If the `maxFeePerBlobGas` or `blobVersionedHashes` are set we treat it as an EIP-4844
+    /// transaction.
+    ///
+    /// Note: Due to the `Default` impl of [`BlockEnv`] (Some(0)) this assumes the `block_blob_fee`
+    /// is always `Some`
+    fn ensure_fees(
+        call_gas_price: Option<U256>,
+        call_max_fee: Option<U256>,
+        call_priority_fee: Option<U256>,
+        block_base_fee: U256,
+        blob_versioned_hashes: Option<&[B256]>,
+        max_fee_per_blob_gas: Option<U256>,
+        block_blob_fee: Option<U256>,
+    ) -> EthResult<Self> {
+        /// Get the effective gas price of a transaction as specfified in EIP-1559 with relevant
+        /// checks.
+        fn get_effective_gas_price(
+            max_fee_per_gas: Option<U256>,
+            max_priority_fee_per_gas: Option<U256>,
+            block_base_fee: U256,
+        ) -> EthResult<U256> {
+            match max_fee_per_gas {
+                Some(max_fee) => {
+                    if max_fee < block_base_fee {
+                        // `base_fee_per_gas` is greater than the `max_fee_per_gas`
+                        return Err(RpcInvalidTransactionError::FeeCapTooLow.into())
+                    }
+                    if max_fee < max_priority_fee_per_gas.unwrap_or(U256::ZERO) {
+                        return Err(
+                            // `max_priority_fee_per_gas` is greater than the `max_fee_per_gas`
+                            RpcInvalidTransactionError::TipAboveFeeCap.into(),
+                        )
+                    }
+                    Ok(min(
+                        max_fee,
+                        block_base_fee
+                            .checked_add(max_priority_fee_per_gas.unwrap_or(U256::ZERO))
+                            .ok_or_else(|| {
+                                EthApiError::from(RpcInvalidTransactionError::TipVeryHigh)
+                            })?,
+                    ))
+                }
+                None => Ok(block_base_fee
+                    .checked_add(max_priority_fee_per_gas.unwrap_or(U256::ZERO))
+                    .ok_or_else(|| EthApiError::from(RpcInvalidTransactionError::TipVeryHigh))?),
+            }
+        }
+
+        let has_blob_hashes =
+            blob_versioned_hashes.as_ref().map(|blobs| !blobs.is_empty()).unwrap_or(false);
+
+        match (call_gas_price, call_max_fee, call_priority_fee, max_fee_per_blob_gas) {
+            (gas_price, None, None, None) => {
+                // either legacy transaction or no fee fields are specified
+                // when no fields are specified, set gas price to zero
+                let gas_price = gas_price.unwrap_or(U256::ZERO);
+                Ok(Self {
+                    gas_price,
+                    max_priority_fee_per_gas: None,
+                    max_fee_per_blob_gas: has_blob_hashes.then_some(block_blob_fee).flatten(),
+                })
+            }
+            (None, max_fee_per_gas, max_priority_fee_per_gas, None) => {
+                // request for eip-1559 transaction
+                let effective_gas_price = get_effective_gas_price(
+                    max_fee_per_gas,
+                    max_priority_fee_per_gas,
+                    block_base_fee,
+                )?;
+                let max_fee_per_blob_gas = has_blob_hashes.then_some(block_blob_fee).flatten();
+
+                Ok(Self {
+                    gas_price: effective_gas_price,
+                    max_priority_fee_per_gas,
+                    max_fee_per_blob_gas,
+                })
+            }
+            (None, max_fee_per_gas, max_priority_fee_per_gas, Some(max_fee_per_blob_gas)) => {
+                // request for eip-4844 transaction
+                let effective_gas_price = get_effective_gas_price(
+                    max_fee_per_gas,
+                    max_priority_fee_per_gas,
+                    block_base_fee,
+                )?;
+                // Ensure blob_hashes are present
+                if !has_blob_hashes {
+                    // Blob transaction but no blob hashes
+                    return Err(RpcInvalidTransactionError::BlobTransactionMissingBlobHashes.into())
+                }
+
+                Ok(Self {
+                    gas_price: effective_gas_price,
+                    max_priority_fee_per_gas,
+                    max_fee_per_blob_gas: Some(max_fee_per_blob_gas),
+                })
+            }
+            _ => {
+                // this fallback covers incompatible combinations of fields
+                Err(EthApiError::ConflictingFeeFieldsInRequest)
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use reth_primitives::constants::GWEI_TO_WEI;
+
+    #[test]
+    fn test_ensure_0_fallback() {
+        let CallFees { gas_price, .. } =
+            CallFees::ensure_fees(None, None, None, U256::from(99), None, None, Some(U256::ZERO))
+                .unwrap();
+        assert!(gas_price.is_zero());
+    }
+
+    #[test]
+    fn test_blob_fees() {
+        let CallFees { gas_price, max_fee_per_blob_gas, .. } =
+            CallFees::ensure_fees(None, None, None, U256::from(99), None, None, Some(U256::ZERO))
+                .unwrap();
+        assert!(gas_price.is_zero());
+        assert_eq!(max_fee_per_blob_gas, None);
+
+        let CallFees { gas_price, max_fee_per_blob_gas, .. } = CallFees::ensure_fees(
+            None,
+            None,
+            None,
+            U256::from(99),
+            Some(&[B256::from(U256::ZERO)]),
+            None,
+            Some(U256::from(99)),
+        )
+        .unwrap();
+        assert!(gas_price.is_zero());
+        assert_eq!(max_fee_per_blob_gas, Some(U256::from(99)));
+    }
+
+    #[test]
+    fn test_eip_1559_fees() {
+        let CallFees { gas_price, .. } = CallFees::ensure_fees(
+            None,
+            Some(U256::from(25 * GWEI_TO_WEI)),
+            Some(U256::from(15 * GWEI_TO_WEI)),
+            U256::from(15 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        )
+        .unwrap();
+        assert_eq!(gas_price, U256::from(25 * GWEI_TO_WEI));
+
+        let CallFees { gas_price, .. } = CallFees::ensure_fees(
+            None,
+            Some(U256::from(25 * GWEI_TO_WEI)),
+            Some(U256::from(5 * GWEI_TO_WEI)),
+            U256::from(15 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        )
+        .unwrap();
+        assert_eq!(gas_price, U256::from(20 * GWEI_TO_WEI));
+
+        let CallFees { gas_price, .. } = CallFees::ensure_fees(
+            None,
+            Some(U256::from(30 * GWEI_TO_WEI)),
+            Some(U256::from(30 * GWEI_TO_WEI)),
+            U256::from(15 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        )
+        .unwrap();
+        assert_eq!(gas_price, U256::from(30 * GWEI_TO_WEI));
+
+        let call_fees = CallFees::ensure_fees(
+            None,
+            Some(U256::from(30 * GWEI_TO_WEI)),
+            Some(U256::from(31 * GWEI_TO_WEI)),
+            U256::from(15 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        );
+        assert!(call_fees.is_err());
+
+        let call_fees = CallFees::ensure_fees(
+            None,
+            Some(U256::from(5 * GWEI_TO_WEI)),
+            Some(U256::from(GWEI_TO_WEI)),
+            U256::from(15 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        );
+        assert!(call_fees.is_err());
+
+        let call_fees = CallFees::ensure_fees(
+            None,
+            Some(U256::MAX),
+            Some(U256::MAX),
+            U256::from(5 * GWEI_TO_WEI),
+            None,
+            None,
+            Some(U256::ZERO),
+        );
+        assert!(call_fees.is_err());
     }
 }
